pytorch/pyproject.toml
[tool.poetry]
name = "pytorch"
version = "0.1.0"
description = ""
authors = ["Siva Chaganti"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.2"
torchsummary = "^1.5.1"
numpy = "^1.26.4"
h5py = "^3.10.0"
pip = "^24.0"
scikit-learn = "^1.4.2"
pillow = "^10.3.0"
torchvision = "^0.17.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
